pyproject.toml
[tool.poetry]
name = "HR-agent"
version = "0.1.0"
description = ""
authors = ["Javier Leguina <javier@leguina.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

faiss-cpu = "^1.7.4"
google-api-python-client = "^2.104.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
langchain = "^0.0.319"
numpy = "^1.26.1"
openai = "^0.28.1"
streamlit = "^1.27.2"
pypdf = "^3.16.4"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"


[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.1"
black = "^23.10.0"
isort = "^5.12.0"
mypy = "^1.6.1"
pre-commit = "^3.5.0"
pylint = "^3.0.1"
types-requests = "^2.31.0.10"
types-urllib3 = "^1.26.25.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
target-version = ['py311']
include = '.*\/.*\.i?py(nb)?$'

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pylint.'MASTER']
ignore = "tests"

[tool.pylint.'MESSAGES CONTROL']
disable = "all"
enable = "C0123, C0200, C0209, E0401, E0602, E1101, E1205, W0611, W0621, W1309, W1404, W1514"
